write plugin configuration into a subdirectory.
the @INCLUDE directive fails if a wildcard does not resolve. since
it is a possible to have a fluentbit pipeline without any filters,
the service would fail to start.
this change assembles the configuration files under a single
directory and uses a single glob directive to load them. fluenbit
reads the files in an alphanumeric order but internally sorts
them correctly (input-filter-output)